[regression] Size the deref bounds tests for LLP64 too
Both new tests assumed long was 8 bytes. On Windows it is 4, so struct c
was 12 bytes and the p + 15 access overran it, while the "oversized"
long read of an int was merely int-sized and stayed in bounds. long long
is 8 bytes under LP64 and LLP64 alike, so both properties now hold on
every target.

+++ regression/esbmc/deref_negative_offset/main.c
@@ -7,7 +7,7 @@ struct c
 {
   char a;
   int b;
-  long d;
+  long long d;
 };
 
 int main(void)
@@ -21,7 +21,7 @@ int main(void)
 
   assert(*(char *)(p + 0) == 1);
   assert(*(int *)(p + 4) == 2);
-  assert(*(long *)(p + 8) == 3);
+  assert(*(long long *)(p + 8) == 3);
   assert(*(char *)(p + 15) == *((char *)&s.d + 7));
 
   return 0;
